Updated prototype for handling json files
