+++ hedtools/hed/util/base_file_input.py
@@ -1,7 +1,7 @@
 import os
 import openpyxl
 import pandas
-from hed.util.hed_event_mapper import EventMapper
+from hed.util.column_mapper import ColumnMapper
 
 class BaseFileInput:
     """Handles parsing the actual on disk hed files to a more general format."""
@@ -25,12 +25,12 @@ class BaseFileInput:
          has_column_names: bool
              True if file has column names. The validation will skip over the first line of the file. False, if
              otherwise.
-         mapper: EventMapper object
-             Pass in a built event mapper(see HedFileInput or EventFileInput for examples), or None to just
+         mapper: ColumnMapper object
+             Pass in a built column mapper(see HedFileInput or EventFileInput for examples), or None to just
              retrieve all columns as hed tags.
          """
         if mapper is None:
-            mapper = EventMapper()
+            mapper = ColumnMapper()
         self._mapper = mapper
         self._filename = filename
         self._worksheet_name = worksheet_name
@@ -79,7 +79,7 @@ class BaseFileInput:
         return self.parse_dataframe()
 
     def iter_raw(self):
-        default_mapper = EventMapper()
+        default_mapper = ColumnMapper()
         return self.parse_dataframe(default_mapper)
 
     def parse_dataframe(self, mapper=None):
